+++ tsc/index.ts
@@ -1,12 +1,67 @@
-import * as usb from './usb';
+import { promisify } from 'util';
 import { WebUSB } from './webusb';
+import { WebUSBDevice } from './webusb/webusb-device';
+import * as usb from './usb';
 
-declare module './usb' {
-    const webusb: WebUSB;
-}
+/**
+ * Convenience method to get the first device with the specified VID and PID, or `undefined` if no such device is present.
+ * @param vid
+ * @param pid
+ */
+const findByIds = (vid: number, pid: number): usb.Device | undefined => {
+    const devices = usb.getDeviceList();
+    return devices.find(item => item.deviceDescriptor.idVendor === vid && item.deviceDescriptor.idProduct === pid);
+};
 
-Object.assign(usb, {
-    webusb: new WebUSB()
-});
+/**
+ * Convenience method to get the device with the specified serial number, or `undefined` if no such device is present.
+ * @param serialNumber
+ */
+const findBySerialNumber = async (serialNumber: string): Promise<usb.Device | undefined> => {
+    const devices = usb.getDeviceList();
 
-export = usb;
+    for (const device of devices) {
+        try {
+            device.open();
+
+            const getStringDescriptor = promisify(device.getStringDescriptor).bind(device);
+            const buffer = await getStringDescriptor(device.deviceDescriptor.iSerialNumber);
+
+            if (buffer && buffer.toString() === serialNumber) {
+                return device;
+            }
+        } catch {
+            // Ignore any errors, device may be a system device or inaccessible
+        } finally {
+            try {
+                device.close();
+            } catch {
+                // Ignore any errors, device may be a system device or inaccessible
+            }
+        }
+    }
+
+    return undefined;
+};
+
+const webusb = new WebUSB();
+const getDeviceList = usb.getDeviceList;
+
+export {
+    // Core usb object for quick access
+    usb,
+
+    // Convenience methods
+    getDeviceList,
+    findByIds,
+    findBySerialNumber,
+
+    // Default WebUSB object (mimics navigator.usb)
+    webusb,
+
+    // WebUSB class for creating custom webusb instances
+    WebUSB,
+
+    // WebUSB Device class for turning a core usb.Device into a webusb device
+    WebUSBDevice
+};
