+++ tsc/usb/index.ts
@@ -1,4 +1,3 @@
-import { promisify } from 'util';
 import { EventEmitter } from 'events';
 import { ExtendedDevice } from './device';
 import * as usb from './bindings';
@@ -39,9 +38,6 @@ declare module './bindings' {
 	function removeAllListeners<K extends keyof DeviceEvents>(event?: K): void;
 	function emit<K extends keyof DeviceEvents>(event: K, arg: DeviceEvents[K]): boolean;
 	function listenerCount<K extends keyof DeviceEvents>(event: K): number;
-
-	function findByIds(vid: number, pid: number): usb.Device | undefined;
-	function findBySerialNumber(serialNumber: string): Promise<usb.Device | undefined>;
 }
 
 usb.on('newListener', event => {
@@ -62,52 +58,6 @@ usb.on('removeListener', event => {
     if (listenerCount === 0) {
         usb._disableHotplugEvents();
     }
-});
-
-/**
- * Convenience method to get the first device with the specified VID and PID, or `undefined` if no such device is present.
- * @param vid
- * @param pid
- */
-const findByIds = (vid: number, pid: number): usb.Device | undefined => {
-    const devices = usb.getDeviceList();
-    return devices.find(item => item.deviceDescriptor.idVendor === vid && item.deviceDescriptor.idProduct === pid);
-};
-
-/**
- * Convenience method to get the device with the specified serial number, or `undefined` if no such device is present.
- * @param serialNumber
- */
-const findBySerialNumber = async (serialNumber: string): Promise<usb.Device | undefined> => {
-    const devices = usb.getDeviceList();
-
-    for (const device of devices) {
-        try {
-            device.open();
-
-            const getStringDescriptor = promisify(device.getStringDescriptor).bind(device);
-            const buffer = await getStringDescriptor(device.deviceDescriptor.iSerialNumber);
-
-            if (buffer && buffer.toString() === serialNumber) {
-                return device;
-            }
-        } catch {
-            // Ignore any errors, device may be a system device or inaccessible
-        } finally {
-            try {
-                device.close();
-            } catch {
-                // Ignore any errors, device may be a system device or inaccessible
-            }
-        }
-    }
-
-    return undefined;
-};
-
-Object.assign(usb, {
-    findByIds,
-    findBySerialNumber
 });
 
 export = usb;
